+++ src/app/[lang]/projects/[id]/party/_components/ButtonCreator.tsx
@@ -1,8 +1,7 @@
 'use client';
 
-import { Grid, Button } from '@mui/material';
-import React from 'react';
-import DialogCreator from './DialogCreator';
+import { Grid, Button, Dialog, DialogActions, DialogContent, DialogContentText, DialogTitle } from '@mui/material';
+import React, { useState } from 'react';
 
 interface ButtonCreatorProps {
   currentUser: { id: string; name: string };
@@ -10,48 +9,87 @@ interface ButtonCreatorProps {
 }
 
 export default function ButtonCreator({ currentUser, currentParty }: ButtonCreatorProps) {
+  const [open, setOpen] = useState(false);
+
   const handleStartProject = () => {
     // Logic to start the project
   };
 
+  const handleCloseRequest = () => {
+    setOpen(true);
+  };
+
   const handleConfirmClose = () => {
     // Logic to close the request
+    setOpen(false);
+  };
+
+  const handleCancelClose = () => {
+    setOpen(false);
   };
 
   return (
     <>
-      {currentUser && currentUser.id && currentUser.id === currentParty.project.creator.id && (
-        <Grid
-          container
-          justifyContent="center"
-          spacing={3}
-        >
-          <Grid item>
-            <DialogCreator onConfirm={handleConfirmClose} />
-          </Grid>
-          <Grid item>
-            <Button
-              variant="contained"
-              sx={{
-                width: '264px',
-                height: '48px',
-                padding: '12px 16px',
-                gap: '16px',
-                borderRadius: '6px 0px 0px 0px',
-                opacity: 1,
-                background: '#18A670',
-                color: '#fff',
-                '&:hover': {
-                  background: '#18A670',
-                },
-              }}
-              onClick={handleStartProject}
-            >
-              Начать проект
-            </Button>
-          </Grid>
+      <Grid
+        container
+        justifyContent="center"
+        spacing={3}
+      >
+        <Grid item>
+          <Button
+            variant="outlined"
+            color={'error'}
+            sx={{
+              width: '264px',
+            }}
+            onClick={handleCloseRequest}
+          >
+            Закрыть запрос
+          </Button>
         </Grid>
-      )}
+        <Grid item>
+          <Button
+            variant="contained"
+            color={'success'}
+            sx={{
+              width: '264px',
+              opacity: 1,
+              background: '#18A670',
+              color: '#fff',
+              '&:hover': {
+                background: '#18A670',
+              },
+            }}
+            onClick={handleStartProject}
+          >
+            Начать проект
+          </Button>
+        </Grid>
+      </Grid>
+      <Dialog
+        open={open}
+        onClose={handleCancelClose}
+      >
+        <DialogTitle>Подтверждение</DialogTitle>
+        <DialogContent>
+          <DialogContentText>Вы уверены, что хотите закрыть запрос?</DialogContentText>
+        </DialogContent>
+        <DialogActions>
+          <Button
+            onClick={handleCancelClose}
+            color="primary"
+          >
+            Отмена
+          </Button>
+          <Button
+            onClick={handleConfirmClose}
+            color="primary"
+            autoFocus
+          >
+            Подтвердить
+          </Button>
+        </DialogActions>
+      </Dialog>
     </>
   );
 }
